Fixed Makefile to give libtool proper version info
Libtool wants 'current:revision:age', where 'current' is the interface
number of the library, 'revision' is the implementation number of that
interface, and 'age' is the number of previous interfaces this interface
is backwards-compatible with.  For normal libraries that only break
backwards compatibility with major version changes, the minor version
number will be used for both 'current' and 'age', and the micro version
number will be used for 'revision'.

For example, for library libhello v1.2.6, the version info passed to
libtool will be '2.6.2' meaning this is the third interface of the
library and it's compatible with the previous two interfaces.  (0 was
the first interface, 1 was the second, and 2 is the third.)  Further-
more it is the seventh implementation of the interface (there have
been six bugs that were fixed without changing the interface).  (The
major version should be part of the name of the library itself, but
since it is 1 we can just leave it off.)  The linker will be able to
use this library for any program that uses this major version number
and it will always use the latest implementation number possible.