Makefile sets libtool age to 0 for unstable series
Doing this because EDO is pretty unstable right now and we need to be
able to break compatibility between minor versions.  (Normally the
libtool age is set to the same number as the minor version to indicate
the library is backwards compatible with all releases within the
same series, i.e. v4.7 is backwards compatible with v4.6, v4.5, etc.)
This should be reset once we release v1.0.  Think this is fine for
the v0.x series.